:running: Move KCP code from 1 file into multiple logical ones
