+++ controlplane/kubeadm/controllers/helpers.go
@@ -1,0 +1,251 @@
+/*
+Copyright 2020 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controllers
+
+import (
+	"context"
+	"strings"
+
+	"github.com/pkg/errors"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	kerrors "k8s.io/apimachinery/pkg/util/errors"
+	"k8s.io/apiserver/pkg/storage/names"
+	clusterv1 "sigs.k8s.io/cluster-api/api/v1alpha3"
+	bootstrapv1 "sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1alpha3"
+	"sigs.k8s.io/cluster-api/controllers/external"
+	controlplanev1 "sigs.k8s.io/cluster-api/controlplane/kubeadm/api/v1alpha3"
+	"sigs.k8s.io/cluster-api/controlplane/kubeadm/internal"
+	"sigs.k8s.io/cluster-api/controlplane/kubeadm/internal/hash"
+	capierrors "sigs.k8s.io/cluster-api/errors"
+	"sigs.k8s.io/cluster-api/util"
+	"sigs.k8s.io/cluster-api/util/kubeconfig"
+	"sigs.k8s.io/cluster-api/util/patch"
+	"sigs.k8s.io/cluster-api/util/secret"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+func (r *KubeadmControlPlaneReconciler) reconcileKubeconfig(ctx context.Context, clusterName client.ObjectKey, endpoint clusterv1.APIEndpoint, kcp *controlplanev1.KubeadmControlPlane) error {
+	if endpoint.IsZero() {
+		return nil
+	}
+
+	_, err := secret.GetFromNamespacedName(ctx, r.Client, clusterName, secret.Kubeconfig)
+	switch {
+	case apierrors.IsNotFound(err):
+		createErr := kubeconfig.CreateSecretWithOwner(
+			ctx,
+			r.Client,
+			clusterName,
+			endpoint.String(),
+			*metav1.NewControllerRef(kcp, controlplanev1.GroupVersion.WithKind("KubeadmControlPlane")),
+		)
+		if createErr != nil {
+			if createErr == kubeconfig.ErrDependentCertificateNotFound {
+				return errors.Wrapf(&capierrors.RequeueAfterError{RequeueAfter: dependentCertRequeueAfter},
+					"could not find secret %q for Cluster %q in namespace %q, requeuing",
+					secret.ClusterCA, clusterName.Name, clusterName.Namespace)
+			}
+			return createErr
+		}
+	case err != nil:
+		return errors.Wrapf(err, "failed to retrieve kubeconfig Secret for Cluster %q in namespace %q", clusterName.Name, clusterName.Namespace)
+	}
+
+	return nil
+}
+
+func (r *KubeadmControlPlaneReconciler) reconcileExternalReference(ctx context.Context, cluster *clusterv1.Cluster, ref corev1.ObjectReference) error {
+	if !strings.HasSuffix(ref.Kind, external.TemplateSuffix) {
+		return nil
+	}
+
+	obj, err := external.Get(ctx, r.Client, &ref, cluster.Namespace)
+	if err != nil {
+		return err
+	}
+
+	// Note: We intentionally do not handle checking for the paused label on an external template reference
+
+	patchHelper, err := patch.NewHelper(obj, r.Client)
+	if err != nil {
+		return err
+	}
+
+	obj.SetOwnerReferences(util.EnsureOwnerRef(obj.GetOwnerReferences(), metav1.OwnerReference{
+		APIVersion: clusterv1.GroupVersion.String(),
+		Kind:       "Cluster",
+		Name:       cluster.Name,
+		UID:        cluster.UID,
+	}))
+
+	return patchHelper.Patch(ctx, obj)
+}
+
+func (r *KubeadmControlPlaneReconciler) cloneConfigsAndGenerateMachine(ctx context.Context, cluster *clusterv1.Cluster, kcp *controlplanev1.KubeadmControlPlane, bootstrapSpec *bootstrapv1.KubeadmConfigSpec, failureDomain *string) error {
+	var errs []error
+
+	// Since the cloned resource should eventually have a controller ref for the Machine, we create an
+	// OwnerReference here without the Controller field set
+	infraCloneOwner := &metav1.OwnerReference{
+		APIVersion: controlplanev1.GroupVersion.String(),
+		Kind:       "KubeadmControlPlane",
+		Name:       kcp.Name,
+		UID:        kcp.UID,
+	}
+
+	// Clone the infrastructure template
+	infraRef, err := external.CloneTemplate(ctx, &external.CloneTemplateInput{
+		Client:      r.Client,
+		TemplateRef: &kcp.Spec.InfrastructureTemplate,
+		Namespace:   kcp.Namespace,
+		OwnerRef:    infraCloneOwner,
+		ClusterName: cluster.Name,
+		Labels:      internal.ControlPlaneLabelsForClusterWithHash(cluster.Name, hash.Compute(&kcp.Spec)),
+	})
+	if err != nil {
+		// Safe to return early here since no resources have been created yet.
+		return errors.Wrap(err, "failed to clone infrastructure template")
+	}
+
+	// Clone the bootstrap configuration
+	bootstrapRef, err := r.generateKubeadmConfig(ctx, kcp, cluster, bootstrapSpec)
+	if err != nil {
+		errs = append(errs, errors.Wrap(err, "failed to generate bootstrap config"))
+	}
+
+	// Only proceed to generating the Machine if we haven't encountered an error
+	if len(errs) == 0 {
+		if err := r.generateMachine(ctx, kcp, cluster, infraRef, bootstrapRef, failureDomain); err != nil {
+			errs = append(errs, errors.Wrap(err, "failed to create Machine"))
+		}
+	}
+
+	// If we encountered any errors, attempt to clean up any dangling resources
+	if len(errs) > 0 {
+		if err := r.cleanupFromGeneration(ctx, infraRef, bootstrapRef); err != nil {
+			errs = append(errs, errors.Wrap(err, "failed to cleanup generated resources"))
+		}
+
+		return kerrors.NewAggregate(errs)
+	}
+
+	return nil
+}
+
+func (r *KubeadmControlPlaneReconciler) cleanupFromGeneration(ctx context.Context, remoteRefs ...*corev1.ObjectReference) error {
+	var errs []error
+
+	for _, ref := range remoteRefs {
+		if ref != nil {
+			config := &unstructured.Unstructured{}
+			config.SetKind(ref.Kind)
+			config.SetAPIVersion(ref.APIVersion)
+			config.SetNamespace(ref.Namespace)
+			config.SetName(ref.Name)
+
+			if err := r.Client.Delete(ctx, config); err != nil && !apierrors.IsNotFound(err) {
+				errs = append(errs, errors.Wrap(err, "failed to cleanup generated resources after error"))
+			}
+		}
+	}
+
+	return kerrors.NewAggregate(errs)
+}
+
+func (r *KubeadmControlPlaneReconciler) generateKubeadmConfig(ctx context.Context, kcp *controlplanev1.KubeadmControlPlane, cluster *clusterv1.Cluster, spec *bootstrapv1.KubeadmConfigSpec) (*corev1.ObjectReference, error) {
+	// Create an owner reference without a controller reference because the owning controller is the machine controller
+	owner := metav1.OwnerReference{
+		APIVersion: controlplanev1.GroupVersion.String(),
+		Kind:       "KubeadmControlPlane",
+		Name:       kcp.Name,
+		UID:        kcp.UID,
+	}
+
+	bootstrapConfig := &bootstrapv1.KubeadmConfig{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:            names.SimpleNameGenerator.GenerateName(kcp.Name + "-"),
+			Namespace:       kcp.Namespace,
+			Labels:          internal.ControlPlaneLabelsForClusterWithHash(cluster.Name, hash.Compute(&kcp.Spec)),
+			OwnerReferences: []metav1.OwnerReference{owner},
+		},
+		Spec: *spec,
+	}
+
+	if err := r.Client.Create(ctx, bootstrapConfig); err != nil {
+		return nil, errors.Wrap(err, "Failed to create bootstrap configuration")
+	}
+
+	bootstrapRef := &corev1.ObjectReference{
+		APIVersion: bootstrapv1.GroupVersion.String(),
+		Kind:       "KubeadmConfig",
+		Name:       bootstrapConfig.GetName(),
+		Namespace:  bootstrapConfig.GetNamespace(),
+		UID:        bootstrapConfig.GetUID(),
+	}
+
+	return bootstrapRef, nil
+}
+
+func (r *KubeadmControlPlaneReconciler) generateMachine(ctx context.Context, kcp *controlplanev1.KubeadmControlPlane, cluster *clusterv1.Cluster, infraRef, bootstrapRef *corev1.ObjectReference, failureDomain *string) error {
+	machine := &clusterv1.Machine{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      names.SimpleNameGenerator.GenerateName(kcp.Name + "-"),
+			Namespace: kcp.Namespace,
+			Labels:    internal.ControlPlaneLabelsForClusterWithHash(cluster.Name, hash.Compute(&kcp.Spec)),
+			OwnerReferences: []metav1.OwnerReference{
+				*metav1.NewControllerRef(kcp, controlplanev1.GroupVersion.WithKind("KubeadmControlPlane")),
+			},
+		},
+		Spec: clusterv1.MachineSpec{
+			ClusterName:       cluster.Name,
+			Version:           &kcp.Spec.Version,
+			InfrastructureRef: *infraRef,
+			Bootstrap: clusterv1.Bootstrap{
+				ConfigRef: bootstrapRef,
+			},
+			FailureDomain: failureDomain,
+		},
+	}
+
+	if err := r.Client.Create(ctx, machine); err != nil {
+		return errors.Wrap(err, "Failed to create machine")
+	}
+	return nil
+}
+
+func (r *KubeadmControlPlaneReconciler) markWithAnnotationKey(ctx context.Context, machine *clusterv1.Machine, annotationKey string) error {
+	if machine == nil {
+		return errors.New("expected machine not nil")
+	}
+	patchHelper, err := patch.NewHelper(machine, r.Client)
+	if err != nil {
+		return errors.Wrapf(err, "failed to create patch helper for machine %s", machine.Name)
+	}
+
+	if machine.Annotations == nil {
+		machine.Annotations = make(map[string]string)
+	}
+	machine.Annotations[annotationKey] = ""
+
+	if err := patchHelper.Patch(ctx, machine); err != nil {
+		return errors.Wrapf(err, "failed to patch machine %s selected for upgrade", machine.Name)
+	}
+	return nil
+}
